+++ memegen/routes/api_root.py
@@ -8,10 +8,10 @@ from ._settings import CHANGES_URL
 from ._utils import route
 
 
-blueprint = Blueprint('root', __name__, url_prefix="/")
+blueprint = Blueprint('root', __name__)
 
 
-@blueprint.route("api")
+@blueprint.route("/api")
 def get():
     """Generate memes from templates."""
     data = OrderedDict()
@@ -24,7 +24,7 @@ def get():
     return data
 
 
-@blueprint.route("CHECK")
+@blueprint.route("/CHECK")
 def handle_checks():
     """Return CHECK_OK for zero-downtime deployment.