+++ tests/test_api_aliases.py
@@ -8,21 +8,21 @@ from .utils import load
 def describe_get():
 
     def it_requires_a_name_to_return_aliases(client):
-        status, data = load(client.get("/aliases/"))
+        status, data = load(client.get("/api/aliases/"))
 
         expect(status) == 200
         expect(data) == []
 
     def it_redirects_with_param(client):
-        status, text = load(client.get("/aliases/?name=foo"), as_json=False)
+        status, text = load(client.get("/api/aliases/?name=foo"), as_json=False)
 
         expect(status) == 302
-        expect(text).contains('<a href="/aliases/foo">')
+        expect(text).contains('<a href="/api/aliases/foo">')
 
     def describe_filter():
 
         def with_single_match(client):
-            status, data = load(client.get("/aliases/sad-biden"))
+            status, data = load(client.get("/api/aliases/sad-biden"))
 
             expect(status) == 200
             expect(data) == {
@@ -32,12 +32,12 @@ def describe_get():
                         'scowl',
                         'window',
                     ],
-                    'template': "http://localhost/templates/sad-biden"
+                    'template': "http://localhost/api/templates/sad-biden"
                 }
             }
 
         def with_many_matches(client):
-            status, data = load(client.get("/aliases/votestakes"))
+            status, data = load(client.get("/api/aliases/votestakes"))
 
             expect(status) == 200
             expect(len(data)) == 5