+++ tests/test_api_links.py
@@ -1,14 +1,16 @@
-# pylint: disable=no-self-use,misplaced-comparison-constant
+# pylint: disable=unused-variable,misplaced-comparison-constant,expression-not-assigned
+
+from expecter import expect
 
 from .conftest import load
 
 
-class TestLinks:
+def describe_get():
 
-    def test_get(self, client):
-        response = client.get("/iw/hello/world")
+    def it_returns_link_options(client):
+        response = client.get("/api/templates/iw/hello/world")
 
-        assert 200 == response.status_code
+        expect(response.status_code) == 200
         assert dict(
             direct=dict(
                 visible="http://localhost/iw/hello/world.jpg",
@@ -20,58 +22,61 @@ class TestLinks:
             ),
         ) == load(response)
 
-    def test_get_with_top_only(self, client):
-        response = client.get("/iw/hello")
+    def with_top_only(client):
+        response = client.get("/api/templates/iw/hello")
 
-        assert 200 == response.status_code
+        expect(response.status_code) == 200
         assert dict(
             visible="http://localhost/iw/hello.jpg",
             masked="http://localhost/_aXcJaGVsbG8J.jpg",
         ) == load(response, key='direct')
 
-    def test_get_with_bottom_only(self, client):
-        response = client.get("/iw/_/hello")
+    def with_bottom_only(client):
+        response = client.get("/api/templates/iw/_/hello")
 
-        assert 200 == response.status_code
+        expect(response.status_code) == 200
         assert dict(
             visible="http://localhost/iw/_/hello.jpg",
             masked="http://localhost/_aXcJXy9oZWxsbwkJ.jpg",
         ) == load(response, key='direct')
 
-    def test_get_redirects_to_dashes(self, client):
-        response = client.get("/iw/HelloThere_World/How-areYOU")
+    def with_dashes(client):
+        response = client.get("/api/templates/iw/HelloThere_World/How-areYOU")
 
-        assert 302 == response.status_code
-        assert '<a href="/iw/hello-there-world/how-are-you">' in \
+        expect(response.status_code) == 302
+        assert '<a href="/api/templates/iw/hello-there-world/how-are-you">' in \
             load(response, as_json=False)
 
-    def test_get_redirects_when_masked(self, client):
+    def when_masked(client):
         response = client.get("/_aXcJaGVsbG8vd29ybGQJ")
 
-        assert 302 == response.status_code
-        assert '<a href="/iw/hello/world">' in load(response, as_json=False)
-
-    def test_get_redirects_when_masked_with_1_line(self, client):
-        response = client.get("/_aXcJaGVsbG8J")
-
-        assert 302 == response.status_code
-        assert '<a href="/iw/hello">' in load(response, as_json=False)
-
-    def test_get_without_extension_redirects_to_template(self, client):
-        response = client.get("/live")
-
-        assert 302 == response.status_code
-        assert '<a href="/templates/live">' in \
+        expect(response.status_code) == 302
+        assert '<a href="/api/templates/iw/hello/world">' in \
             load(response, as_json=False)
 
-    def test_get_redirects_when_alias(self, client):
-        response = client.get("/insanity-wolf")
+    def when_masked_with_1_line(client):
+        response = client.get("/_aXcJaGVsbG8J")
+
+        expect(response.status_code) == 302
+        assert '<a href="/api/templates/iw/hello">' in \
+            load(response, as_json=False)
+
+    def when_alias(client):
+        response = client.get("/api/templates/insanity-wolf")
+
+        expect(response.status_code) == 302
+        assert '<a href="/api/templates/iw">' in load(response, as_json=False)
+
+    def when_alias_with_text(client):
+        response = client.get("/api/templates/insanity-wolf/hello/world")
+
+        expect(response.status_code) == 302
+        assert '<a href="/api/templates/iw/hello/world">' in \
+            load(response, as_json=False)
+
+    def old_api_is_still_supported_via_redirect(client):
+        response = client.get("/iw/top/bottom")
 
         assert 302 == response.status_code
-        assert '<a href="/templates/iw">' in load(response, as_json=False)
-
-    def test_get_redirects_when_alias_with_text(self, client):
-        response = client.get("/insanity-wolf/hello/world")
-
-        assert 302 == response.status_code
-        assert '<a href="/iw/hello/world">' in load(response, as_json=False)
+        assert '<a href="/api/templates/iw/top/bottom">' in \
+            load(response, as_json=False)