+++ tests/test_api_templates.py
@@ -7,8 +7,8 @@ from .conftest import load
 
 def describe_get():
 
-    def when_default_text(client):
-        response = client.get("/templates/iw")
+    def it_returns_template_info(client):
+        response = client.get("/api/templates/iw")
 
         assert 200 == response.status_code
         assert dict(
@@ -16,52 +16,52 @@ def describe_get():
             description="http://knowyourmeme.com/memes/insanity-wolf",
             aliases=['insanity', 'insanity-wolf', 'iw'],
             styles=[],
-            example="http://localhost/iw/does-testing/in-production",
+            example="http://localhost/api/templates/iw/does-testing/in-production",
         ) == load(response)
 
     def when_no_default_text(client):
-        response = client.get("/templates/keanu")
+        response = client.get("/api/templates/keanu")
 
         assert 200 == response.status_code
-        assert "http://localhost/keanu/your-text/goes-here" == \
+        assert "http://localhost/api/templates/keanu/your-text/goes-here" == \
             load(response)['example']
 
     def when_alternate_sytles_available(client):
-        response = client.get("/templates/sad-biden")
+        response = client.get("/api/templates/sad-biden")
 
         assert 200 == response.status_code
         assert ['down', 'scowl', 'window'] == load(response)['styles']
 
     def when_dashes_in_key(client):
-        response = client.get("/templates/awkward-awesome")
+        response = client.get("/api/templates/awkward-awesome")
 
         assert 200 == response.status_code
 
     def it_returns_list_when_no_key(client):
-        response = client.get("/templates/")
+        response = client.get("/api/templates/")
 
         assert 200 == response.status_code
         data = load(response)
-        assert "http://localhost/templates/iw" == data['Insanity Wolf']
+        assert "http://localhost/api/templates/iw" == data['Insanity Wolf']
         assert len(data) >= 20  # there should be many memes
 
-    def it_redirects_when_text_is_provided(client):
-        response = client.get("/templates/iw/top/bottom")
-
-        assert 302 == response.status_code
-        assert '<a href="/iw/top/bottom">' in load(response, as_json=False)
-
     def it_redirects_when_key_is_an_alias(client):
-        response = client.get("/templates/insanity-wolf")
+        response = client.get("/api/templates/insanity-wolf")
 
         assert 302 == response.status_code
-        assert '<a href="/templates/iw">' in load(response, as_json=False)
+        assert '<a href="/api/templates/iw">' in load(response, as_json=False)
+
+    def old_api_is_still_supported_via_redirect(client):
+        response = client.get("/templates/")
+
+        assert 302 == response.status_code
+        assert '<a href="/api/templates/">' in load(response, as_json=False)
 
 
 def describe_post():
 
     def new_templates_are_created_using_github(client):
-        response = client.post("/templates/")
+        response = client.post("/api/templates/")
 
         assert 403 == response.status_code
         assert dict(
@@ -70,7 +70,7 @@ def describe_post():
 
     def can_create_a_new_meme(client):
         params = {'top': "foo", 'bottom': "bar"}
-        response = client.post("/templates/fry", data=params)
+        response = client.post("/api/templates/fry", data=params)
 
         expect(response.status_code) == 303
         expect(load(response, as_json=False)).contains(